Move delete all button
