+++ src/pages/preferences/preferencesTabs/notificationTab/utils.tsx
@@ -1,3 +1,5 @@
+import { useMemo } from "react";
+import { useQuery } from "@apollo/client";
 import styled from "@emotion/styled";
 import { LeafyGreenTableRow } from "@leafygreen-ui/table/new";
 import {
@@ -7,10 +9,50 @@ import {
   getVersionRoute,
 } from "constants/routes";
 import { size } from "constants/tokens";
-import { GeneralSubscription, Selector } from "gql/generated/types";
+import {
+  GeneralSubscription,
+  Selector,
+  UserSubscriptionsQuery,
+  UserSubscriptionsQueryVariables,
+} from "gql/generated/types";
+import { USER_SUBSCRIPTIONS } from "gql/queries";
 import { ResourceType } from "types/triggers";
 
-export const processSubscriptionData = (
+export const useSubscriptionData = () => {
+  const { data } = useQuery<
+    UserSubscriptionsQuery,
+    UserSubscriptionsQueryVariables
+  >(USER_SUBSCRIPTIONS);
+
+  const globalSubscriptionIds = useMemo(() => {
+    const {
+      buildBreakId,
+      commitQueueId,
+      patchFinishId,
+      patchFirstFailureId,
+      spawnHostExpirationId,
+      spawnHostOutcomeId,
+    } = data?.userSettings?.notifications ?? {};
+    return new Set([
+      buildBreakId,
+      commitQueueId,
+      patchFinishId,
+      patchFirstFailureId,
+      spawnHostExpirationId,
+      spawnHostOutcomeId,
+    ]);
+  }, [data?.userSettings?.notifications]);
+
+  const subscriptions = useMemo(
+    () =>
+      processSubscriptionData(data?.user?.subscriptions, globalSubscriptionIds),
+    [data?.user?.subscriptions, globalSubscriptionIds]
+  );
+
+  return subscriptions;
+};
+
+const processSubscriptionData = (
   subscriptions: GeneralSubscription[],
   globalSubscriptionIds: Set<string>
 ) => {
